+++ src/main/java/rocks/cleancode/conventionalcommit/ConventionalCommit.java
@@ -4,7 +4,16 @@ import java.util.Map;
 import java.util.Optional;
 import java.util.function.Predicate;
 
+import static java.lang.String.format;
+import static java.util.stream.Collectors.joining;
+
 public class ConventionalCommit {
+
+    public static final String EMPTY_STRING = "";
+
+    public static final String NEWLINE = format("%n");
+
+    public static final String DOUBLE_NEWLINE = NEWLINE + NEWLINE;
 
     private final String type;
 
@@ -19,12 +28,12 @@ public class ConventionalCommit {
     private final Map<String, String> footer;
 
     public ConventionalCommit(
-            String type,
-            String scope,
-            boolean exclamation,
-            String description,
-            String body,
-            Map<String, String> footer
+        String type,
+        String scope,
+        boolean exclamation,
+        String description,
+        String body,
+        Map<String, String> footer
     ) {
         this.type = type;
         this.scope = scope;
@@ -52,7 +61,7 @@ public class ConventionalCommit {
 
     public Optional<String> body() {
         return Optional.ofNullable(body)
-                .filter(this::isNotBlank);
+            .filter(this::isNotBlank);
     }
 
     public Map<String, String> footer() {
@@ -65,13 +74,62 @@ public class ConventionalCommit {
 
     private boolean isNotBlank(String string) {
         return Optional.ofNullable(string)
-                .map(String::trim)
-                .filter(isEmpty().negate())
-                .isPresent();
+            .map(String::trim)
+            .filter(isEmpty().negate())
+            .isPresent();
     }
 
     public boolean breakingChange() {
         return exclamation || footer.containsKey("BREAKING CHANGE");
+    }
+
+    @Override
+    public String toString() {
+        return format(
+                "%s%s%s: %s%s%s",
+                type,
+                scopeToString(),
+                exclamationToString(),
+                description,
+                bodyToString(),
+                footerToString()
+            )
+            .trim();
+    }
+
+    private String scopeToString() {
+        return scope()
+            .map(this::surroundScope)
+            .orElse(EMPTY_STRING);
+    }
+
+    private String exclamationToString() {
+        if (exclamation) {
+            return "!";
+        }
+
+        return EMPTY_STRING;
+    }
+
+    private String bodyToString() {
+        return body()
+            .map(DOUBLE_NEWLINE::concat)
+            .orElse(EMPTY_STRING);
+    }
+
+    private String footerToString() {
+        return footer().entrySet()
+            .stream()
+            .map(this::joinFooterEntry)
+            .collect(joining(NEWLINE, DOUBLE_NEWLINE, EMPTY_STRING));
+    }
+
+    private String joinFooterEntry(Map.Entry<String, String> entry) {
+        return format("%s%s%s", entry.getKey(), ": ", entry.getValue());
+    }
+
+    private String surroundScope(String scope) {
+        return format("(%s)", scope);
     }
 
 }