+++ CHANGELOG.md
@@ -5,5 +5,6 @@
 ### Features
 
 - Conventional commit parser
+- Conventional commit generator
 
 [Unreleased]: https://github.com/clean-code-rocks/conventional-commit/commits/main